Remove builtin scripts (#9)
Summary: Removes `example.lua` and `panda.lua` from the builtin scripts as we are no longer going to package the actual templates within the CLI. This also involves updating the `RuntimeScript` loading functionality to not look for a `builtin:...` prefix. Additionally, as a result, remove the `makepdf script` subcommand as we no longer have a need for it.

Test Plan: Observe that the `makepdf script` subcommand is gone and that provide `--script builtin:panda` no longer succeeds.