Refactor PdfColor lua representation into table
Summary: Previously, PdfColor was represented as a string. Now, we represent as a table with `red`, `green`, and `blue` fields that are byte values between 0 and 255. The table implements a `__tostring` metatable method, and there is also now a `pdf.utils.color()` function to help with converting.

Definitions have been updated to reference a `pdf.common.ColorLike`.

Test Plan: `cargo test`